requirements.txt
Quart>=0.19.6
azure-eventgrid==4.11.0
aiohttp>= 3.11.9
azure-communication-callautomation==1.4.0b1
openai
